From Wim: ChangeLog entries for pyroot.diff:
 o) return proxy on access through class (needed for help() on public data)
 o) added "string" to ROOT.std as a special STL case
 o) allow unpacking of std::pair in for-loops over a container
 o) from Scott: fixed leak of argv in TPyROOTApplication and
       of PyMethodDefs in Utility::AddToClass
 o) port to p2.5b3 and backwards compatibility updates

There are a few places where I could not get all to compile with both p2.5
and older pythons, so I resorted to CPP #defines.

There are two changes of note in p2.5 from an extension module POV: 64 bit
indexing support, and module loading has been cleaned up on the Mac (meaning,
CINT will no longer see duplicate dictionaries from python). In addition,
there is a new module 'ctypes' that allows programmatic support to functions
in a shared library; I still have to see whether anything useful can be
culled out of that.


git-svn-id: http://root.cern.ch/svn/root/trunk@15994 27541ba8-7e3a-0410-8455-c3a389f83636

+++ pyroot/src/Converters.h
@@ -1,4 +1,4 @@
-// @(#)root/pyroot:$Name:  $:$Id: Converters.h,v 1.17 2005/12/03 04:00:15 pcanal Exp $
+// @(#)root/pyroot:$Name:  $:$Id: Converters.h,v 1.18 2005/12/07 06:16:16 brun Exp $
 // Author: Wim Lavrijsen, Jan 2005
 #ifndef PYROOT_CONVERTERS_H
 #define PYROOT_CONVERTERS_H
@@ -59,7 +59,7 @@ namespace PyROOT {
       virtual PyObject* FromMemory( void* );                                  \
       virtual Bool_t ToMemory( PyObject*, void* );                            \
    private:                                                                   \
-     int fSize;                                                               \
+      Py_ssize_t fSize;                                                       \
    }
 
 // converters for built-ins
