+++ pyroot/src/PyBufferFactory.h
@@ -1,4 +1,4 @@
-// @(#)root/pyroot:$Name:  $:$Id: PyBufferFactory.h,v 1.6 2005/05/25 06:23:36 brun Exp $
+// @(#)root/pyroot:$Name:  $:$Id: PyBufferFactory.h,v 1.7 2005/09/09 05:19:10 brun Exp $
 // Author: Wim Lavrijsen, Apr 2004
 
 #ifndef PYROOT_TPYBUFFERFACTORY_H
@@ -17,21 +17,21 @@ class TPyBufferFactory {
 public:
    static TPyBufferFactory* Instance();
 
-   PyObject* PyBuffer_FromMemory( Short_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( Short_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( Short_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( UShort_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( UShort_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( UShort_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( Int_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( Int_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( Int_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( UInt_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( UInt_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( UInt_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( Long_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( Long_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( Long_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( ULong_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( ULong_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( ULong_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( Float_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( Float_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( Float_t* buf, PyObject* sizeCallback );
-   PyObject* PyBuffer_FromMemory( Double_t* buf, int size = -1 );
+   PyObject* PyBuffer_FromMemory( Double_t* buf, Py_ssize_t size = -1 );
    PyObject* PyBuffer_FromMemory( Double_t* buf, PyObject* sizeCallback );
 
 protected:
